requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
geoalchemy2==0.15.2
celery==5.4.0
redis==5.0.7
httpx==0.27.2
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
structlog==24.1.0
flower==2.0.1
